+++ ESCenter.Client.Application/Contracts/Courses/Dtos/CourseCreateForLearnerDto.cs
@@ -1,7 +1,5 @@
 ﻿using ESCenter.Domain.Aggregates.Courses;
 using ESCenter.Domain.Aggregates.Subjects.ValueObjects;
-using ESCenter.Domain.Aggregates.Users.ValueObjects;
-using ESCenter.Domain.Shared;
 using ESCenter.Domain.Shared.Courses;
 using FluentValidation;
 using Mapster;
@@ -12,8 +10,8 @@ public class CourseCreateForLearnerDto
 {
     public string Title { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public LearningMode LearningMode { get; set; } 
-    public float Fee { get; set; }
+    public LearningMode LearningMode { get; set; }
+    public decimal Fee { get; set; }
     public Gender GenderRequirement { get; set; }
     public AcademicLevel AcademicLevelRequirement { get; set; }
     public Gender LearnerGender { get; set; }
@@ -75,7 +73,7 @@ public class CourseForLearnerCreateDtoValidator : AbstractValidator<CourseCreate
             .WithMessage("Learning mode must be a valid option.");
 
         RuleFor(x => x.Fee)
-            .InclusiveBetween(0, float.MaxValue)
+            .InclusiveBetween(0, decimal.MaxValue)
             .WithMessage("Fee must be a non-negative number.");
 
         RuleFor(x => x.GenderRequirement)
@@ -106,10 +104,6 @@ public class CourseForLearnerCreateDtoValidator : AbstractValidator<CourseCreate
             .Matches(@"^\d{10}$") // Assuming 10-digit phone number
             .WithMessage("Contact number must be 10 digits long.");
 
-        // RuleFor(x => x.LearnerId)
-        //     .NotEmpty()
-        //     .WithMessage("Learner ID is required.");
-
         RuleFor(x => x.MinutePerSession)
             .InclusiveBetween(1, 180)
             .WithMessage("Minutes per session must be between 1 and 180.");
@@ -125,10 +119,5 @@ public class CourseForLearnerCreateDtoValidator : AbstractValidator<CourseCreate
         RuleFor(x => x.SubjectId)
             .GreaterThan(0)
             .WithMessage("Subject ID must be a positive number.");
-
-        // RuleFor(x => x.SubjectName)
-        //     .NotEmpty()
-        //     .MaximumLength(100)
-        //     .WithMessage("Subject name must be less than 100 characters long.");
     }
 }
